requirements.txt
smbus-cffi
Adafruit_DHT
statsd
six